= Fixed miracle machine's psi-field being based entirely around the psi-helmet.
Note: Unlike the Brain Scorcher, who's psi-field logic is based around whether or not it has been turned off, the Miracle Machine's logic lynchpins around whether the player has a psi-helmet in their inventory.

By also disabling it if the primary switch is disabled, we can turn off this psi-field too if story mode is disabled just like we can with the brain scorcher.